New page for scheduled jobs overview and new
