+++ app/routes/overview/scheduled-jobs.js
@@ -1,3 +1,17 @@
 import Route from '@ember/routing/route';
+import DataTableRouteMixin from 'ember-data-table/mixins/route';
+import { service } from '@ember/service';
 
-export default class OverviewScheduledJobsRoute extends Route {}
+export default class OverviewScheduledJobsRoute extends Route.extend(
+  DataTableRouteMixin
+) {
+  @service store;
+
+  modelName = 'scheduled-job';
+
+  mergeQueryOptions(param) {
+    return {
+      sort: param.sort,
+    };
+  }
+}
