+++ app/routes/scheduled-job2/details.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class ScheduledJob2DetailsRoute extends Route {}
