+++ app/templates/overview/jobs.hbs
@@ -45,28 +45,23 @@
       <AuDataTableThSortable
         @field="created"
         @currentSorting={{this.sort}}
-        @label="Created"
-        @class="data-table__header-title"/>
+        @label="Created"/>
       <AuDataTableThSortable
         @field="modified"
         @currentSorting={{this.sort}}
-        @label="Modified"
-        @class="data-table__header-title"/>
+        @label="Modified"/>
       <AuDataTableThSortable
         @field="creator"
         @currentSorting={{this.sort}}
-        @label="Creator"
-        @class="data-table__header-title"/>
+        @label="Creator"/>
       <AuDataTableThSortable
         @field="operation"
         @currentSorting={{this.sort}}
-        @label="Operation"
-        @class="data-table__header-title"/>
+        @label="Operation"/>
       <AuDataTableThSortable
         @field="status"
         @currentSorting={{this.sort}}
-        @label="Status"
-        @class="data-table__header-title"/>
+        @label="Status"/>
       <th>Actions</th>
     </c.header>
     <c.body as |row|>
