+++ app/templates/scheduled-job2/details.hbs
@@ -1,0 +1,3 @@
+{{page-title "Scheduled Job Details"}}
+
+{{outlet}}
