--- app/templates/scheduled-jobs.hbs
@@ -1,1 +1,0 @@
-{{outlet}}
