--- app/templates/scheduled-jobs/index.hbs
@@ -1,68 +1,0 @@
-<AuToolbar
-  @border="bottom"
-  @size="large"
-  as |Group|>
-  <Group>
-    <AuHeading @skin="2">Scheduled jobs overview</AuHeading>
-  </Group>
-  <Group>
-    <div class="au-u-text-right">
-      <LinkTo @route="scheduled-jobs.new" class="au-c-button" type="button">Create new scheduled job</LinkTo>
-    </div>
-  </Group>
-</AuToolbar>
-
-<AuMainContainer as |m|>
-  <m.sidebar>
-    <SidebarMenu />
-  </m.sidebar>
-  <m.content @scroll={{false}}>
-    <AuDataTable @content={{this.model}} @fields="title description" @noDataMessage="No jobs found" @sort={{this.sort}}
-      @size={{this.size}} @page={{this.page}} as |t|>
-      <t.content as |c|>
-        <c.header>
-          <th class="au-u-visible-small-up">uri</th>
-          <AuDataTableThSortable @field="created" @currentSorting={{this.sort}} @label="Created"
-            @class="data-table__header-title" />
-          <AuDataTableThSortable @field="modified" @currentSorting={{this.sort}} @label="Modified"
-            @class="data-table__header-title" />
-          <AuDataTableThSortable @field="operation" @currentSorting={{this.sort}} @label="Operation"
-            @class="data-table__header-title" />
-          <AuDataTableThSortable @field="schedule" @currentSorting={{this.sort}} @label="Schedule"
-            @class="data-table__header-title" />
-          <AuDataTableThSortable @field="title" @currentSorting={{this.sort}} @label="Title"
-            @class="data-table__header-title" />
-          <th class="au-u-visible-small-up">Action</th>
-        </c.header>
-        <c.body as |row|>
-          <td>
-            {{row.uri}}
-          </td>
-          <td>
-            {{format-date row.created}}
-          </td>
-          <td class="au-u-visible-small-up">
-            {{format-date row.modified}}
-          </td>
-          <td class="au-u-visible-small-up au-u-word-break">
-            {{operation-name row.operation}}
-          </td>
-          <td class="au-u-visible-small-up">
-            {{row.schedule.repeatFrequency}}
-          </td>
-          <td class="au-u-visible-medium-up">
-            {{row.title}}
-          </td>
-          <td class="au-u-visible-large-up">
-            <LinkTo @route="scheduled-jobs.details" @model={{row}}>
-              <AuButton @skin="primary">
-                view
-              </AuButton>
-            </LinkTo>
-          </td>
-        </c.body>
-      </t.content>
-    </AuDataTable>
-
-  </m.content>
-</AuMainContainer>
